Move responsibility of growth curve into SPManager.
