+++ nestkernel/synaptic_element.cpp
@@ -60,8 +60,7 @@ nest::SynapticElement::SynapticElement( const SynapticElement& se )
   , growth_rate_( se.growth_rate_ )
   , tau_vacant_( se.tau_vacant_ )
 {
-  growth_curve_ =
-    kernel().connection_builder_manager.new_growth_curve( se.growth_curve_->get_name() );
+  growth_curve_ = kernel().sp_manager.new_growth_curve( se.growth_curve_->get_name() );
   assert( growth_curve_ != 0 );
   DictionaryDatum gc_parameters = DictionaryDatum( new Dictionary );
   se.get( gc_parameters );
@@ -73,8 +72,7 @@ nest::SynapticElement& nest::SynapticElement::operator=( const SynapticElement& 
   if ( this != &other )
   {
     // 1: allocate new memory and copy the elements
-    GrowthCurve* new_gc =
-      kernel().connection_builder_manager.new_growth_curve( other.growth_curve_->get_name() );
+    GrowthCurve* new_gc = kernel().sp_manager.new_growth_curve( other.growth_curve_->get_name() );
     DictionaryDatum gc_parameters = DictionaryDatum( new Dictionary );
 
     other.get( gc_parameters );
@@ -129,7 +127,7 @@ nest::SynapticElement::set( const DictionaryDatum& d )
     Name growth_curve_name( getValue< std::string >( d, names::growth_curve ) );
     if ( not growth_curve_->is( growth_curve_name ) )
     {
-      growth_curve_ = kernel().connection_builder_manager.new_growth_curve( growth_curve_name );
+      growth_curve_ = kernel().sp_manager.new_growth_curve( growth_curve_name );
     }
   }
   growth_curve_->set( d );
