+++ testsuite/regressiontests/issue-105.sli
@@ -162,7 +162,7 @@ M_ERROR setverbosity
   
   /min_delay 0.81 def  % rounded to 0.8
   /max_delay 1.49 def  % rounded to 1.5
-  
+
   0 << /min_delay min_delay /max_delay max_delay >> SetStatus
   
   /n /iaf_neuron Create def
